+++ packages/emoji/src/constants.ts
@@ -1,3 +1,5 @@
+// We hardcode these values so that newer emojis
+// are not displayed for devices that do not support
+// them yet. Support cycle is twice a year.
 export const MAX_EMOJI_VERSION = 13.1;
-
 export const LATEST_DATASET_VERSION = '6.1.0';